+++ CHANGELOG.md
@@ -7,11 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 
 ## [Unreleased]
 
-### Changed
+## Changed
 
 - `api`, `recorder`, `player`
   - Bump dependencies to the latest available versions.
   - Update babel config to make use of the latest babel features.
+  - Updated all changelog files to point at this specific file. This way all changes can be logged
+    and be visible at one location.
 
 ## [v4.1.0] - 2019-03-27