+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 ## Changed
 
 - `api`, `recorder`, `player`
+  - Bump dependencies to the latest available versions.
+  - Update babel config to make use of the latest babel features.
   - Updated all changelog files to point at this specific file. This way all changes can be logged
     and be visible at one location.