+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 
 ## [Unreleased]
 
+## [v4.2.0] - 2019-04-17
+
 ### Added
 
 - `recorder`
@@ -106,7 +108,8 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 - Improve README.md documentation.
 - Changed the getUserAuth and getOAuth2Token to use the new API auth functions.
 
-[unreleased]: https://github.com/itslanguage/itslanguage-js/compare/v4.1.0...HEAD
+[unreleased]: https://github.com/itslanguage/itslanguage-js/compare/v4.2.0...HEAD
+[v4.2.0]: https://github.com/itslanguage/itslanguage-js/compare/v4.1.0...v4.2.0
 [v4.1.0]: https://github.com/itslanguage/itslanguage-js/compare/v4.0.2...v4.1.0
 [v4.0.2]: https://github.com/itslanguage/itslanguage-js/compare/v4.0.1...v4.0.2
 [v4.0.1]: https://github.com/itslanguage/itslanguage-js/compare/v4.0.0...v4.0.1